requirements.txt
# Dictionary OCR dependencies
# PDF handling
pymupdf>=1.24.0
Pillow>=10.0.0

# Vision LLM APIs
google-generativeai>=0.8.0
openai>=1.50.0
anthropic>=0.40.0

# Open-source OCR
surya-ocr>=0.8.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
